publishing_api/requirements/base.txt
Django>=5.1,<7.0
PyYAML>=6.0
requests>=2.31
python-dotenv>=1.0
